Refactor the whole thing
The file .bashrc_synced file gets sourced from the original .profile
file, and all other .*_synced files are sourced through it.

Files not ending with _synced have to be either linked or sourced
in their main counterparts contained in the home directory.

The script dotinit.sh got updated so that it solves some of this.

+++ dotinit.sh
@@ -1,22 +1,58 @@
 #!/bin/bash
 
-touch ~/.bashrc_local
-touch ~/.bash_aliases_local
-touch ~/.bash_functions_local
+echo "0. Source the entry point in ~/.profile:
 
-# ln -sv ~/dotfiles/.bashrc ~
+if [ -n "$BASH_VERSION" ]; then
+    # INSERT THIS PART
+    # include dotfiles repo entry point
+    if [ -f "$HOME/dotfiles/bashrc_synced" ]; then
+	. "$HOME/dotfiles/bashrc_synced"
+    fi
+    # / INSERT THIS PART
 
-ln -sv ~/dotfiles/.bash_aliases ~
-ln -sv ~/dotfiles/.bash_functions ~
+    # include .bashrc if it exists
+    if [ -f "$HOME/.bashrc" ]; then
+	. "$HOME/.bashrc"
+    fi
+fi
 
-ln -sv ~/dotfiles/.gitignore_global ~
-ln -sv ~/dotfiles/.tmux.conf ~
+Insert before the regular .bashrc to make local overrides possible.
 
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-echo "Enter 'tmux' and CTRL+b SHIFT+i to install plugins."
-echo "Source dotfiles in ~/.bashrc according to need."
+
+"
+
+# GIT STUFF
+if [ -f "$HOME/.gitignore_global" ]; then
+	echo ".gitignore_global not linked up because it already exists"
+else
+    ln -sv ~/dotfiles/.gitignore_global ~
+fi
 
 git config --global core.excludesfile ~/.gitignore_global
 git config --global core.editor "nano"
 git config --global alias.s status
+# / GIT STUFF
 
+# TMUX STUFF
+if [ ! -f "$HOME/.tmux.conf" ]; then
+    echo "$HOME/.tmux.conf did not exist, created"
+    touch "$HOME/.tmux.conf"
+fi
+
+if ! grep -q "source-file ~/dotfiles/.tmux.conf_synced" "$HOME/.tmux.conf"
+then
+    echo "" >> "$HOME/.tmux.conf"
+    echo "# content before include will be overridden by the common settings" >> "$HOME/.tmux.conf"
+    echo "source-file ~/dotfiles/.tmux.conf_synced" >> "$HOME/.tmux.conf"
+fi
+
+if [ ! -f "$HOME/.tmux/plugins/tpm" ]; then
+    git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+fi
+
+echo "1. Enter 'tmux' and CTRL+b SHIFT+i to install plugins."
+# / TMUX STUFF
+
+
+ln -sv ~/dotfiles/.inputrc ~
+ln -sv ~/dotfiles/.prompt ~
